Make UIListselect typesafe for reference and pointer types. Now it is possible to use it without all those ugly casts (the casts are encapsulated in the implementation). Get rid of many magic numbers in the code (button IDs).
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@1701 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/game_options_menu.h
@@ -47,19 +47,21 @@ private:
 	UIButton   save_game;
 	UIButton   load_game;
 	UIButton   exit_game;
-	
+
 	/** Returns the horizontal/vertical spacing between buttons. */
 	uint hspacing() const {return 5;};
 	uint vspacing() const {return 5;};
-	
+
 	/** Returns the horizontal/vertical margin between edge and buttons. */
 	uint hmargin() const {return 2 * hspacing();}
 	uint vmargin() const {return 2 * vspacing();}
-	
+
 	/** Returns the width of a button in a row with nr_buttons buttons. */
-	uint buttonw(const uint nr_buttons) const
-	{return (get_inner_w() * 1.45 - (nr_buttons + 3) * hspacing()) / nr_buttons;}
-	
+	uint buttonw(const uint nr_buttons) const {
+		return static_cast<const uint>
+			((get_inner_w() * 1.45 - (nr_buttons + 3) * hspacing()) / nr_buttons);
+	}
+
 	/**
 	 * Returns the x coordinate of the (left edge of) button number nr in a row
 	 * with nr_buttons buttons.
